change data type on default fact measure setup from decimal 38 10 to BIGINT. close #10

+++ factTableTemplate/factFinalSchema/MetaObjName.sql
@@ -3,5 +3,5 @@
 	 [SK_Date] INT NOT NULL CONSTRAINT [df_factFinalSchema_MetaObjName_SK_Date] DEFAULT (-1) CONSTRAINT [FK_factFinalSchema_MetaObjName_SK_Date_To_Date] FOREIGN KEY REFERENCES [dimensionFinalSchema].[Date_dim],	/*SurrogateKey_ReplacementPoint|Date| */
 	 [SK_Date_RP_RolePlay] INT NOT NULL CONSTRAINT [df_factFinalSchema_MetaObjName_SK_Date_RP_RolePlay] DEFAULT (-1) CONSTRAINT [FK_factFinalSchema_MetaObjName_SK_Date_To_Date_RP_RolePlay] FOREIGN KEY REFERENCES [dimensionRpFinalSchema].[Date_dim_RolePlay],	/*SurrogateKey_RolePlay_ReplacementPoint|Date|RolePlay| */
 	 [ImportantDegenerateDimension_OnFactDim] NVARCHAR(500) NOT NULL CONSTRAINT [df_factFinalSchema_MetaObjName_ImportantDegenerateDimension_OnFactDim] DEFAULT ('Context Not Specified'),	/*DegenerateDimensionAttribute_ReplacementPoint|ImportantDegenerateDimension_OnFactDim| */
-	 [ImportantFact] DECIMAL(38,10) NULL	/*DimensionAttribute_ReplacementPoint|ImportantFact|,*/
+	 [ImportantFact] BIGINT NULL	/*DimensionAttribute_ReplacementPoint|ImportantFact|,*/
 )
